Remove all warnings.

+++ animalfarm/send.cpp
@@ -14,7 +14,7 @@ bool ipc_send_ack() {
 	if (!ipc_send_data_to_port(&ACK, dwToWrite)) {
 		return FALSE;
 	}
-	int ms = duration_cast<milliseconds>(
+	long long ms = duration_cast<milliseconds>(
 		system_clock::now().time_since_epoch()
 		).count() - 1480980000000;
 	LOGMESSAGE(L"Send ACK--------" << ms << "\n");
@@ -27,7 +27,7 @@ bool ipc_send_enq() {
 	if (!ipc_send_data_to_port(&ENQ, dwToWrite)) {
 		return FALSE;
 	}
-	int ms = duration_cast<milliseconds>(
+	long long ms = duration_cast<milliseconds>(
 		system_clock::now().time_since_epoch()
 		).count() - 1480980000000;
 	LOGMESSAGE(L"Send ENQ--------" << ms << "\n");
@@ -39,7 +39,7 @@ bool ipc_send_packet(const char* packet) {
 	if (!ipc_send_data_to_port(packet, dwToWrite)) {
 		return FALSE;
 	}
-	int ms = duration_cast<milliseconds>(
+	long long ms = duration_cast<milliseconds>(
 		system_clock::now().time_since_epoch()
 		).count() - 1480980000000;
 	LOGMESSAGE(L"Send PACKET-----" << ms <<"\n");
